Switch all packet stats / counters to uint64

+++ capture/afpacket/afpacket/afpacket_test.go
@@ -192,7 +192,7 @@ func TestPipe(t *testing.T) {
 		require.Nil(t, <-errChan)
 		stats, err := mockSrc.Stats()
 		require.Nil(t, err)
-		require.Equal(t, capture.Stats{PacketsReceived: int(n * n)}, stats)
+		require.Equal(t, capture.Stats{PacketsReceived: uint64(n * n)}, stats)
 		require.Nil(t, mockSrc.Close())
 	}()
 
@@ -219,7 +219,7 @@ func TestPipe(t *testing.T) {
 	require.Nil(t, <-errChan2)
 	stats, err := mockSrc2.Stats()
 	require.Nil(t, err)
-	require.Equal(t, capture.Stats{PacketsReceived: int(n * n)}, stats)
+	require.Equal(t, capture.Stats{PacketsReceived: uint64(n * n)}, stats)
 	require.Nil(t, mockSrc2.Close())
 }
 
@@ -342,7 +342,7 @@ func testCaptureMethods(t *testing.T, fn func(t *testing.T, src *MockSource, i, 
 	// Evaluate packet statistics
 	stats, err := mockSrc.Stats()
 	require.Nil(t, err)
-	require.Equal(t, capture.Stats{PacketsReceived: int(n * n)}, stats)
+	require.Equal(t, capture.Stats{PacketsReceived: uint64(n * n)}, stats)
 
 	// Close the mock source
 	require.Nil(t, mockSrc.Close())
